refactor(test): mv metrics test to a separate file
